pyproject.toml
[tool.poetry]
name = "clara-ai"
version = "0.0.4"
description = "CLARA: Code Language Assistant & Repository Analyzer"
authors = ["Cristóbal Carnero Liñán <cristobal@seednapse.ai>"]
readme = "README.md"
packages = [{include = "clara"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = {version = ">=0.0.139", extras = ["llms"]}
fire = "^0.5.0"
chromadb = "^0.3.21"
rich = "^13.3.3"
tiktoken = "^0.3.3"
prompt-toolkit = "^3.0.38"
click = "^8.1.3"
pyyaml = "^6.0"
mergedeep = "^1.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
clara = "clara.cli:main"
